pawpal_plus_ai_assistant/requirements.txt
streamlit>=1.30
pytest>=7.0
scikit-learn>=1.4
google-genai>=1.0.0
python-dotenv>=1.0.0
